+++ csatar-plugins/csatar/models/MandateType.php
@@ -8,6 +8,8 @@ use Csatar\Csatar\Models\Troop;
 use Input;
 use Lang;
 use Model;
+use October\Rain\Database\Collection;
+use Session;
 
 /**
  * Model
@@ -86,11 +88,11 @@ class MandateType extends Model
     function getOrganizationTypeModelNameOptions()
     {
         return [
-            Association::getOrganizationTypeModelName() => Association::getOrganizationTypeModelNameUserFriendly(),
-            District::getOrganizationTypeModelName() => District::getOrganizationTypeModelNameUserFriendly(),
-            Patrol::getOrganizationTypeModelName() => Patrol::getOrganizationTypeModelNameUserFriendly(),
-            Team::getOrganizationTypeModelName() => Team::getOrganizationTypeModelNameUserFriendly(),
-            Troop::getOrganizationTypeModelName() => Troop::getOrganizationTypeModelNameUserFriendly(),
+            Association::getModelName() => Association::getOrganizationTypeModelNameUserFriendly(),
+            District::getModelName() => District::getOrganizationTypeModelNameUserFriendly(),
+            Patrol::getModelName() => Patrol::getOrganizationTypeModelNameUserFriendly(),
+            Team::getModelName() => Team::getOrganizationTypeModelNameUserFriendly(),
+            Troop::getModelName() => Troop::getOrganizationTypeModelNameUserFriendly(),
         ];
     }
 
@@ -127,15 +129,15 @@ class MandateType extends Model
         else {
             $inputData = Input::get('data');
             if ($inputData && array_key_exists('association', $inputData) && !empty($inputData['association'])) {
-                $mandate_model_type = District::getOrganizationTypeModelName();
+                $mandate_model_type = District::getModelName();
                 $association_id = $inputData['association'];
             }
             else if ($inputData && array_key_exists('district', $inputData) && !empty($inputData['district'])) {
-                $mandate_model_type = Team::getOrganizationTypeModelName();
+                $mandate_model_type = Team::getModelName();
                 $association_id = District::find($inputData['district'])->getAssociationId();
             }
             else if ($inputData && array_key_exists('team', $inputData) && !empty($inputData['team'])) {
-                $mandate_model_type = array_key_exists('troop', $inputData) ? Patrol::getOrganizationTypeModelName() : Troop::getOrganizationTypeModelName();
+                $mandate_model_type = array_key_exists('troop', $inputData) ? Patrol::getModelName() : Troop::getModelName();
                 $association_id = Team::find($inputData['team'])->getAssociationId();
             }
         }
@@ -143,21 +145,79 @@ class MandateType extends Model
         return $association_id && $mandate_model_type ? $query->where('association_id', $association_id)->where('organization_type_model_name', $mandate_model_type) : $query->whereNull('id');
     }
 
-    function scopeMandateTypeIdsInAssociation($query, $associationId) {
+    public static function getAllMandateTypeIdsInAssociation($associationId)
+    {
         return self::where('association_id', $associationId)->get()->pluck('id');
     }
 
-    function scopeScoutMandateTypeIdInAssociation($query, $associationId): array {
+    public static function getScoutMandateTypeIdInAssociation($associationId): array
+    {
         $scoutMandateType = self::where('association_id', $associationId)
             ->where('organization_type_model_name', '\Csatar\Csatar\Models\Scout')
             ->first();
         return $scoutMandateType ? [ $scoutMandateType->id ] : [];
     }
 
-    function scopeGuestMandateTypeInAssociation($query, $associationId): array {
-        $guestMandateType = self::where('association_id', $associationId)
+    public static function getGuestMandateTypeIdInAssociation($associationId): ?int
+    {
+        $sessionRecord = Session::get('guest.mandateTypeIds');
+
+        if(!empty($sessionRecord) && $sessionRecordForAssociation = $sessionRecord->where('associationId', $associationId)->first()) {
+            return $sessionRecordForAssociation['guestMandateTypeId'];
+        }
+
+        if(empty($sessionRecord)){
+            $sessionRecord = new Collection([]);
+        }
+
+        $guestMandateTypeId = self::where('association_id', $associationId)
             ->where('organization_type_model_name', self::MODEL_NAME_GUEST)
-            ->first();
-        return $guestMandateType ? [ $guestMandateType->id ] : [];
+            ->first()->id;
+
+        $sessionRecord = $sessionRecord->replace([ $associationId => [
+            'associationId' => $associationId,
+            'savedToSession' => date('Y-m-d H:i'),
+            'guestMandateTypeId'=> $guestMandateTypeId,
+        ]]);
+
+        Session::put('guest.mandateTypeIds', $sessionRecord);
+
+        return $guestMandateTypeId;
     }
+
+    public function getMandateTypeOptions($scopes = null){
+        if (!empty($scopes['association']->value)) {
+            return MandateType::whereIn('association_id', array_keys($scopes['association']->value))
+                ->lists('name', 'id')
+                ;
+        }
+        else {
+            return MandateType::orderBy('name', 'asc')->lists('association_id', 'id');
+        }
+    }
+
+    public function getModelOptions(){
+        return MandateType::distinct()->where('organization_type_model_name', '<>', self::MODEL_NAME_GUEST)->orderBy('organization_type_model_name', 'asc')->lists('organization_type_model_name', 'organization_type_model_name');
+    }
+
+    public static function getMandatesTypesForMatrix () {
+        $associationIds = Association::all()->pluck('id');
+
+        $mandateTypes = [];
+
+        foreach ($associationIds as $associationId) {
+            $mandatesTypesInAssociation = self::where('association_id', $associationId)->orderBy('nest_left', 'desc')->get();
+            $mandateTypes[$associationId] = $mandatesTypesInAssociation->map(function ($item){
+                return [
+                    'id'                            => $item->id,
+                    'name'                          => $item->name,
+                    'joinAsName'                    => str_replace('-', '_', str_slug($item->name)),
+                    'organization_type_model_name'  => $item->organization_type_model_name
+                ];
+            });
+        }
+
+        return $mandateTypes;
+    }
+
 }