+++ csatar-plugins/csatar/models/Patrol.php
@@ -25,7 +25,7 @@ class Patrol extends OrganizationBase
         'facebook_page' => 'url|regex:(facebook)|nullable',
         'logo' => 'image|nullable',
         'age_group' => 'required',
-         'team' => 'required',
+        'team' => 'required',
     ];
 
     /**
@@ -42,7 +42,7 @@ class Patrol extends OrganizationBase
         if ($this->troop_id && $this->troop->team->id != $this->team_id) {
             throw new \ValidationException(['troop' => Lang::get('csatar.csatar::lang.plugin.admin.patrol.troopNotInTheTeamError')]);
         }
-       
+
         // check that the required mandates are set for now
         $this->validateRequiredMandates($this->attributes);
     }
@@ -73,7 +73,6 @@ class Patrol extends OrganizationBase
         'team_id',
         'troop_id',
         'logo',
-        'age_group_id'
     ];
 
     /**
@@ -81,13 +80,21 @@ class Patrol extends OrganizationBase
      */
 
     public $belongsTo = [
-        'team' => '\Csatar\Csatar\Models\Team',
+        'team' => [
+            '\Csatar\Csatar\Models\Team',
+            'formBuilder' => [
+                'requiredBeforeRender' => true,
+            ],
+        ],
         'troop' => '\Csatar\Csatar\Models\Troop',
         'age_group' => '\Csatar\Csatar\Models\AgeGroup',
     ];
 
     public $hasMany = [
-        'scouts' => '\Csatar\Csatar\Models\Scout',
+        'scouts' => [
+            '\Csatar\Csatar\Models\Scout',
+            'label' => 'csatar.csatar::lang.plugin.admin.scout.scouts',
+        ],
         'mandates' => [
             '\Csatar\Csatar\Models\Mandate',
             'key' => 'mandate_model_id',
@@ -118,7 +125,11 @@ class Patrol extends OrganizationBase
     }
 
     public $morphOne = [
-        'content_page' => ['\Csatar\Csatar\Models\ContentPage', 'name' => 'model']
+        'content_page' => [
+            '\Csatar\Csatar\Models\ContentPage',
+            'name' => 'model',
+            'label' => 'csatar.csatar::lang.plugin.admin.general.contentPage',
+        ],
     ];
 
     /**
@@ -149,7 +160,7 @@ class Patrol extends OrganizationBase
         }
         return [];
     }
-    
+
     /**
      * Return all patrols, which belong to the given team
      */